moved a couple more files
