--- backend/src/config/README.md
@@ -1,3 +1,0 @@
-# config
-
-Configuration files (e.g., env variables)
